browser-use-agent/pyproject.toml
[project]
name = "browser-use-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "browser-use>=0.12.6",
    "langchain-anthropic>=1.3.1",
    "python-dotenv>=1.2.1",
]
